apps/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.6
python-dotenv>=1.0
loguru>=0.7
httpx>=0.27
langchain>=0.2
langgraph>=0.2
twilio
python-multipart
google-generativeai
